Maps Category: Dark Mode
Updates samples in Maps category to support dark mode.
