+++ arcgis-runtime-samples-macos/Assets.xcassets/Discard.imageset/Contents.json
@@ -2,22 +2,25 @@
   "images" : [
     {
       "idiom" : "universal",
-      "scale" : "1x",
-      "filename" : "iOS8_TabBar_Discard22.png"
+      "filename" : "iOS8_TabBar_Discard22.png",
+      "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "scale" : "2x",
-      "filename" : "iOS8_TabBar_Discard44.png"
+      "filename" : "iOS8_TabBar_Discard44.png",
+      "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "scale" : "3x",
-      "filename" : "iOS8_TabBar_Discard66.png"
+      "filename" : "iOS8_TabBar_Discard66.png",
+      "scale" : "3x"
     }
   ],
   "info" : {
     "version" : 1,
     "author" : "xcode"
+  },
+  "properties" : {
+    "template-rendering-intent" : "template"
   }
 }
